app/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
more-itertools==9.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.1
pytest-django==4.5.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.14.0
gunicorn==20.1.0
django-environ==0.9.0